+++ app/src/main/java/ru/subnak/easybike/domain/model/Journey.kt
@@ -8,7 +8,7 @@ data class Journey(
     val userID: Int,
     val date: Long,
     val speed: Int,
-    val distance: Double,
+    val distance: Float,
     val duration: Long,
     val img: String,
     val journeyValues: List<JourneyValue>,
